*** empty log message ***

+++ bundles/org.eclipse.swt/Eclipse SWT PI/cocoa/library/os.h
@@ -20,7 +20,7 @@
 //#include "os_custom.h"
 
 #ifndef __i386__
-#define objc_msgSend_fpret (int)objc_msgSend
+#define objc_msgSend_fpret objc_msgSend
 #endif
 
 #endif /* INC_os_H */
